examples/int8/training/segmentation/requirements.txt
allbumentations
tqdm
pillow
numpy